requirements.txt
requests==2.25.1
Django==3.2.4
beautifulsoup4==4.9.3
Pillow==8.3.1
